Add a public interface to display Lottie Animations directly from a CALayer
